+++ gradle.properties
@@ -10,7 +10,7 @@ loader_version=0.14.24
 
 # Mod Properties
 mod_version=1.0.0
-maven_group=name.modid
+maven_group=bmartin127.advanced-peripherals-remastered
 archives_base_name=advancedperipherals
