+++ src/main/java/bmartin127/advanced-peripherals-remastered/AdvancedPeripheralsConfig.java
@@ -1,4 +1,4 @@
-package name.modid;
+package bmartin127.advanced-peripherals-remastered;
 
 import com.github.imifou.jsonschema.module.addon.annotation.JsonSchema;
