+++ src/main/java/bmartin127/advanced-peripherals-remastered/common/blocks/GeoScannerBlock.java
@@ -1,7 +1,7 @@
-package name.modid.common.blocks;
+package bmartin127.advanced-peripherals-remastered.common.blocks;
 
 
-import name.modid.common.blocks.blockentities.GeoScannerBlockEntity;
+import bmartin127.advanced-peripherals-remastered.common.blocks.blockentities.GeoScannerBlockEntity;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
