+++ src/main/java/bmartin127/advanced-peripherals-remastered/common/item/pocket/PocketGeoScannerUpgrade.java
@@ -1,9 +1,9 @@
-package name.modid.common.item.pocket;
+package bmartin127.advanced-peripherals-remastered.common.item.pocket;
 
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 import dan200.computercraft.api.pocket.IPocketAccess;
-import name.modid.common.peripherals.GeoScannerPeripheral;
+import bmartin127.advanced-peripherals-remastered.common.peripherals.GeoScannerPeripheral;
 import dan200.computercraft.ComputerCraftAPIImpl;
 
 public class PocketGeoScannerUpgrade extends BasePocketUpgrade{
