+++ src/main/java/bmartin127/advanced-peripherals-remastered/common/utl/StringUtil.java
@@ -1,0 +1,9 @@
+package bmartin127.advanced-peripherals-remastered.common.utl;
+
+import static bmartin127.advanced-peripherals-remastered.Advancedperipherals.MODID;
+
+public class StringUtil {
+    public static String pocket(String name) {
+        return String.format("pocket.%s.%s", MODID, name);
+    }
+}
