+++ src/main/java/bmartin127/advanced-peripherals-remastered/integrations/cc/AdvancedPeripheralsBlockEntityTypes.java
@@ -1,13 +1,13 @@
-package name.modid.integrations.cc;
+package bmartin127.advanced-peripherals-remastered.integrations.cc;
 
-import name.modid.common.blocks.blockentities.GeoScannerBlockEntity;
+import bmartin127.advanced-peripherals-remastered.common.blocks.blockentities.GeoScannerBlockEntity;
 import net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityTypeBuilder;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 
-import static name.modid.Advancedperipherals.MODID;
-import static name.modid.common.setup.Blocks.GEO_SCANNER;
+import static bmartin127.advanced-peripherals-remastered.Advancedperipherals.MODID;
+import static bmartin127.advanced-peripherals-remastered.common.setup.Blocks.GEO_SCANNER;
 
 public class AdvancedPeripheralsBlockEntityTypes {
     public static final BlockEntityType<GeoScannerBlockEntity> GEO_SCANNER_BLOCK_ENTITY = Registry.register(
