+++ src/main/java/bmartin127/advanced-peripherals-remastered/integrations/cc/AdvancedPeripheralsPeripheralProviders.java
@@ -1,4 +1,4 @@
-package name.modid.integrations.cc;
+package bmartin127.advanced-peripherals-remastered.integrations.cc;
 
 
 import dan200.computercraft.ComputerCraft;
@@ -11,12 +11,12 @@ import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.BlockState;
-import name.modid.AdvancedPeripheralsConfig;
-import name.modid.common.blocks.blockentities.GeoScannerBlockEntity;
-import name.modid.common.peripherals.GeoScannerPeripheral;
+import bmartin127.advanced-peripherals-remastered.AdvancedPeripheralsConfig;
+import bmartin127.advanced-peripherals-remastered.common.blocks.blockentities.GeoScannerBlockEntity;
+import bmartin127.advanced-peripherals-remastered.common.peripherals.GeoScannerPeripheral;
 
 
-import name.modid.common.setup.Blocks;
+import bmartin127.advanced-peripherals-remastered.common.setup.Blocks;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
