--- src/main/java/name/modid/common/utl/StringUtil.java
@@ -1,9 +1,0 @@
-package name.modid.common.utl;
-
-import static name.modid.Advancedperipherals.MODID;
-
-public class StringUtil {
-    public static String pocket(String name) {
-        return String.format("pocket.%s.%s", MODID, name);
-    }
-}
