+++ src/main/resources/advancedperipherals.mixins.json
@@ -1,6 +1,6 @@
 {
 	"required": true,
-	"package": "name.modid.mixin",
+	"package": "bmartin127.advanced-peripherals-remastered.mixin",
 	"compatibilityLevel": "JAVA_17",
 	"mixins": [
 		"ExampleMixin"
